requirements.txt
discord.py==2.0.1
python-dotenv==0.21.0
pyTelegramBotAPI==4.7.0
